requirements.txt
py-cord==2.6.0
pytz==2024.1
selenium==4.24.0
chromedriver-binary-auto==0.3.1
python-dotenv==1.0.1